requirements.txt
discord.py==2.6.4
Pillow
aiohttp
python-dotenv
psutil
